agent/requirements.txt
# Core FastAPI and async dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# PydanticAI and AI dependencies
pydantic-ai==0.0.13
pydantic==2.5.0
openai==1.3.0

# Database dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# WebSocket support
websockets==12.0

# Additional utilities
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1